TCES-12 refactor(auth) refactor middleware to make testing easier
