+++ server/src/middlewares/auth/isAdmin.js
@@ -1,0 +1,18 @@
+const User = require('../../models/user.model');
+
+const isAdmin = async (req, res, next) => {
+    // First check the session, then check the database
+    if (!req.user) {
+        return res.status(403).json({status: "error", message: "You are not logged in"});
+    }
+    if (!req.user.is_admin) {
+        return res.status(403).json({status: "error", message: "You are not an admin"});
+    }
+    const user = await User.findOne({ where: { email: req.user.username } });
+    if (!user || !user.is_admin) {
+        return res.status(403).json({status: "error", message: "You are not logged in"});
+    }
+    return next();
+}
+
+module.exports = isAdmin;
